+++ Libplanet.Net/Messages/Message.cs
@@ -145,10 +145,10 @@ namespace Libplanet.Net.Messages
         public AppProtocolVersion Version { get; set; }
 
         /// <summary>
-        /// The sender <see cref="Peer"/> of the message.
+        /// The sender <see cref="BoundPeer"/> of the message.
         /// </summary>
         [LogAsScalar]
-        public Peer? Remote { get; set; }
+        public BoundPeer? Remote { get; set; }
 
         /// <summary>
         /// The timestamp of the message is created.