+++ CHANGES.md
@@ -66,6 +66,8 @@ To be released.
     returns null.  [[#915], [#2200]]
  -  (Libplanet.Net) `ITransport.AsPeer` and `Swarm<T>.AsPeer` type changed from
     `Peer` to `BoundPeer`.  [[#2215]]
+ -  (Libplanet.Net) All public return type, parameter type, and property type
+    of `Peer` changed to `BoundPeer`.   [[#2228]]
 
 ### Backward-incompatible network protocol changes
 
@@ -135,12 +137,13 @@ To be released.
 
 [#915]: https://github.com/planetarium/libplanet/issues/915
 [#1538]: https://github.com/planetarium/libplanet/issues/1538
-[#2187]:https://github.com/planetarium/libplanet/issues/2187
+[#2187]: https://github.com/planetarium/libplanet/issues/2187
 [#2200]: https://github.com/planetarium/libplanet/pull/2200
 [#2201]: https://github.com/planetarium/libplanet/pull/2201
 [#2215]: https://github.com/planetarium/libplanet/pull/2215
 [#2216]: https://github.com/planetarium/libplanet/pull/2216
 [#2219]: https://github.com/planetarium/libplanet/pull/2219
+[#2228]: https://github.com/planetarium/libplanet/pull/2228
 [#2232]: https://github.com/planetarium/libplanet/pull/2231